Rename vaccination row to be a component
